[#80108] 근접측정소 API 데이터  삽입 완료

+++ src/main/java/com/example/realpilot/excelModel/RegionData.java
@@ -18,7 +18,7 @@ public class RegionData {
     private String createdDate;
     private Integer gridX;
     private Integer gridY;
-    private Double tmX;
-    private Double tmY;*/
+    private Float tmX;
+    private Float tmY;*/
 
 }
